requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.51.0
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langchain-text-splitters==0.3.0
chromadb==0.5.7
pypdf==4.3.1
pydantic==2.9.2
textblob==0.20.0
